moved generated grammar files to sub-folder
